Final code for Cython threading (float version)

+++ src/11-cython/perf/math_core.pyx
@@ -2,9 +2,9 @@ from libc.math cimport sqrt
 
 import cython
 
-def do_math(start: cython.int = 0, num: cython.int = 10):
-    pos: cython.int = start
-    k_sq: cython.int = 1000 * 1000
+def do_math(start: cython.float = 0, num: cython.float = 10):
+    pos: cython.float = start
+    k_sq: cython.float = 1000 * 1000
 
     with nogil:
         while pos < num:
